+++ src/common/middleware/logger.middleware.go
@@ -1,13 +1,13 @@
 package middleware
 
 import (
-	"log"
+	"log/slog"
 	"time"
 
 	"github.com/gofiber/fiber/v3"
 )
 
-func Logger() fiber.Handler {
+func Logger(log *slog.Logger) fiber.Handler {
 	return func(c fiber.Ctx) error {
 		start := time.Now()
 
@@ -18,7 +18,12 @@ func Logger() fiber.Handler {
 		method := c.Method()
 		path := c.Path()
 
-		log.Printf("[%s] %d | %10v | %s", method, status, latency, path)
+		log.Info("http_request",
+			slog.String("method", method),
+			slog.Int("status", status),
+			slog.Duration("latency", latency),
+			slog.String("path", path),
+		)
 		return err
 	}
 }